Address the 3 items deferred from round 20's report
- Removed the unbounded readString() overloads from BufferedFile.h and
  Serialization.h: dead code with no production caller (only a test used
  the BufferedFile.h one), sitting right next to the correctly-bounded
  tryReadString() as a live foot-gun for any future code that reaches for
  it on a corrupted .bin cache file, reintroducing the unbounded-resize()
  heap-abort class fixed elsewhere this session. Switched the one test call
  site to tryReadString().
- TocNcxParser::currentDepth / TocNavParser::olDepth were uint8_t counters
  incremented once per nested <navPoint>/<ol> with no cap, wrapping past
  255 levels of nesting in a crafted EPUB nav document and throwing off the
  zero-check each depends on to know when nesting has fully unwound. Widened
  both to uint32_t so the open/close balance stays correct at any depth, and
  clamp only at the point the value is handed to
  BookMetadataCache::createTocEntry()'s uint8_t `level` parameter -- that
  parameter is a display-only indent count, so saturating it at 255 is a
  cosmetic difference on a pathological input, not a functional one.
- Removed a dead sanity check in SdCardFont::prewarmStyle(): dataLength is
  a uint16_t (max 65535), so `dataLength > 65536` can never be true --  the
  compiler already flags this as tautological. It provided no actual
  protection; the real overflow guard is the running-sum check a few lines
  below, which is self-sufficient and doesn't depend on this per-glyph
  check. Updated that check's comment, which had referenced the removed one.

+++ lib/Epub/Epub/parsers/TocNavParser.h
@@ -25,8 +25,15 @@ class TocNavParser final : public Print {
   ParserState state = START;
   BookMetadataCache* cache;
 
-  // Track nesting depth for <ol> elements to determine TOC depth
-  uint8_t olDepth = 0;
+  // Track nesting depth for <ol> elements to determine TOC depth. Wide
+  // enough that open/close balance stays correct no matter how deep a
+  // crafted nav document nests -- a uint8_t here would wrap past 255
+  // levels, throwing off the zero-check this depends on to know when
+  // nesting has fully unwound. The value actually stored
+  // (BookMetadataCache::createTocEntry's `level` parameter) is clamped
+  // separately to fit its own uint8_t; this counter's only job is tracking
+  // depth correctly, not fitting in a byte.
+  uint32_t olDepth = 0;
   // Current entry data being collected
   std::string currentLabel;
   std::string currentHref;
